+++ app/build.gradle
@@ -51,29 +51,24 @@ android {
 
 dependencies {
 
-    implementation 'androidx.core:core-ktx:1.13.0'
+    implementation 'androidx.core:core-ktx:1.8.0'
     implementation platform('org.jetbrains.kotlin:kotlin-bom:1.8.0')
-    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.7.0'
-    implementation 'androidx.activity:activity-compose:1.9.0'
+    implementation 'androidx.lifecycle:lifecycle-runtime-ktx:2.3.1'
+    implementation 'androidx.activity:activity-compose:1.5.1'
     implementation platform('androidx.compose:compose-bom:2022.10.00')
     implementation 'androidx.compose.ui:ui'
     implementation 'androidx.compose.ui:ui-graphics'
     implementation 'androidx.compose.ui:ui-tooling-preview'
-    implementation 'androidx.compose.material3:material3'
-    implementation platform('androidx.compose:compose-bom:2023.08.00')
     implementation 'androidx.constraintlayout:constraintlayout-compose:1.0.1'
-    implementation platform('androidx.compose:compose-bom:2023.08.00')
+    implementation 'androidx.compose.material3:material3-android:1.2.1'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'Gradle.androidx.test.espresso:espresso-core:3.5.1@aar'
-    implementation "androidx.compose.material:material-icons-extended:1.6.6"
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
     androidTestImplementation platform('androidx.compose:compose-bom:2022.10.00')
     androidTestImplementation 'androidx.compose.ui:ui-test-junit4'
-    androidTestImplementation platform('androidx.compose:compose-bom:2023.08.00')
-    androidTestImplementation platform('androidx.compose:compose-bom:2023.08.00')
     debugImplementation 'androidx.compose.ui:ui-tooling'
     debugImplementation 'androidx.compose.ui:ui-test-manifest'
+    implementation "androidx.compose.material:material-icons-extended:1.6.6"
 
     //Splash Api
     implementation "androidx.core:core-splashscreen:1.0.1"
@@ -106,7 +101,8 @@ dependencies {
 
     //Paging 3
     def paging_version = "3.1.1"
-    implementation "androidx.paging:paging-compose:3.2.1"
+    implementation "androidx.paging:paging-runtime:$paging_version"
+    implementation "androidx.paging:paging-compose:3.2.0-rc01"
 
     //Room
     def room_version = "2.5.2"
