+++ app/src/main/AndroidManifest.xml
@@ -10,13 +10,13 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/App.Starting.Theme"
-        tools:targetApi="31">
+        android:theme="@style/Theme.MoodyApp"
+        tools:targetApi="34">
         <activity
-            android:name=".Login"
-            android:exported="false"
+            android:name=".presentation.login.Login"
+            android:exported="true"
             android:label="@string/title_activity_login"
-            android:theme="@style/App.Starting.Theme">
+            android:theme="@style/Theme.MoodyApp">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -27,7 +27,7 @@
             android:name=".MainActivity"
             android:exported="false"
             android:label="@string/app_name"
-            android:theme="@style/App.Starting.Theme"></activity>
+            android:theme="@style/Theme.MoodyApp"></activity>
     </application>
 
 </manifest>
