+++ app/src/main/java/com/example/moodyapp/presentation/onboarding/components/OnBoardingPage.kt
@@ -5,10 +5,12 @@ import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -29,28 +31,32 @@ fun OnBoardingPage(
     modifier: Modifier = Modifier,
     page: Page
 ) {
-    Column(modifier = modifier) {
-        Image(
-            modifier = Modifier
-                .fillMaxWidth()
-                .fillMaxHeight(fraction = 0.6f),
-            painter = painterResource(id = page.image),
-            contentDescription = null,
-            contentScale = ContentScale.Crop
-        )
-        Spacer(modifier = Modifier.height(MediumPadding1))
-        Text(
-            text = page.title,
-            modifier = Modifier.padding(horizontal = MediumPadding2),
-            style = MaterialTheme.typography.displaySmall.copy(fontWeight = FontWeight.Bold),
-            color = colorResource(id = R.color.display_small)
-        )
-        Text(
-            text = page.description,
-            modifier = Modifier.padding(horizontal = MediumPadding2),
-            style = MaterialTheme.typography.bodyMedium,
-            color = colorResource(id = R.color.text_medium)
-        )
+    Surface(
+        modifier = Modifier.fillMaxSize(), color = MaterialTheme.colorScheme.surface
+    ) {
+        Column(modifier = modifier) {
+            Image(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .fillMaxHeight(fraction = 0.6f),
+                painter = painterResource(id = page.image),
+                contentDescription = null,
+                contentScale = ContentScale.Crop
+            )
+            Spacer(modifier = Modifier.height(MediumPadding1))
+            Text(
+                text = page.title,
+                modifier = Modifier.padding(horizontal = MediumPadding2),
+                style = MaterialTheme.typography.displaySmall.copy(fontWeight = FontWeight.Bold),
+                color = colorResource(id = R.color.display_small)
+            )
+            Text(
+                text = page.description,
+                modifier = Modifier.padding(horizontal = MediumPadding2),
+                style = MaterialTheme.typography.bodyMedium,
+                color = colorResource(id = R.color.text_medium)
+            )
+        }
     }
 }
