+++ app/src/main/java/com/example/moodyapp/ui/theme/Type.kt
@@ -45,5 +45,11 @@ val Typography = Typography(
         fontFamily = Quicksand,
         fontWeight = FontWeight.Normal,
         lineHeight = 19.sp
-    )
+    ),
+    labelMedium = TextStyle(
+        fontSize = 18.sp,
+        fontFamily = Quicksand,
+        fontWeight = FontWeight.Normal,
+        lineHeight = 19.sp
+    ),
 )
